fix(refactor): move UploadZone into a separated file
